requirements.txt
nltk==3.7
numpy==1.22.4
pandas==1.5.1
Pillow==9.3.0
requests==2.28.1
streamlit==1.14.0
streamlit_lottie==0.0.3
